+++ CHANGELOG.md
@@ -13,12 +13,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [0.3.0] - 2025-07-17
+
+### Added
+
+- Complete game page logic (human mode)
+
 ## [0.2.1] - 2025-07-11
 
 ### Added
 
 - Add initial match data to Pinia store
-- Complete round page logic
+- Complete round start page logic
 
 ### Changed